+++ packages/protobuf/src/services/cosmos-ibc-core.ts
@@ -1,4 +1,8 @@
 export { Query as IbcChannelService } from '../../gen/ibc/core/channel/v1/query_connect.js';
+export { Msg as IbcChannelMsgService } from '../../gen/ibc/core/channel/v1/tx_connect.js';
+
 export { Query as IbcClientService } from '../../gen/ibc/core/client/v1/query_connect.js';
 export { Msg as IbcClientMsgService } from '../../gen/ibc/core/client/v1/tx_connect.js';
+
 export { Query as IbcConnectionService } from '../../gen/ibc/core/connection/v1/query_connect.js';
+export { Msg as IbcConnectionMsgService } from '../../gen/ibc/core/connection/v1/tx_connect.js';