+++ packages/ui/src/WalletBalance/index.tsx
@@ -1,5 +1,5 @@
 import type { MouseEventHandler } from 'react';
-import styled, { type DefaultTheme } from 'styled-components';
+import { styled, type DefaultTheme } from 'styled-components';
 import { Wallet } from 'lucide-react';
 import type { BalancesResponse } from '@penumbra-zone/protobuf/penumbra/view/v1/view_pb';
 import { getFormattedAmtFromValueView } from '@penumbra-zone/types/value-view';