+++ packages/perspective/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @penumbra-zone/perspective
 
+## 31.0.0
+
+### Patch Changes
+
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+  - @penumbra-zone/protobuf@6.1.0
+  - @penumbra-zone/bech32m@8.0.0
+  - @penumbra-zone/wasm@29.0.0
+  - @penumbra-zone/getters@18.0.0
+
 ## 30.0.0
 
 ### Patch Changes