+++ packages/services/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @penumbra-zone/router
 
+## 35.0.0
+
+### Patch Changes
+
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+  - @penumbra-zone/storage@31.0.0
+  - @penumbra-zone/protobuf@6.1.0
+  - @penumbra-zone/bech32m@8.0.0
+  - @penumbra-zone/wasm@29.0.0
+  - @penumbra-zone/query@32.0.0
+  - @penumbra-zone/types@24.0.0
+  - @penumbra-zone/getters@18.0.0
+  - @penumbra-zone/crypto-web@25.0.0
+
+## 34.0.1
+
+### Patch Changes
+
+- Updated dependencies [e343d22]
+- Updated dependencies [f9b8c07]
+- Updated dependencies [e343d22]
+- Updated dependencies [e343d22]
+  - @penumbra-zone/storage@30.0.1
+  - @penumbra-zone/query@31.0.1
+
 ## 34.0.0
 
 ### Patch Changes