+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @penumbra-zone/storage
 
+## 31.0.0
+
+### Patch Changes
+
+- b6e32f8: Bump registry version
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+  - @penumbra-zone/protobuf@6.1.0
+  - @penumbra-zone/bech32m@8.0.0
+  - @penumbra-zone/wasm@29.0.0
+  - @penumbra-zone/types@24.0.0
+  - @penumbra-zone/getters@18.0.0
+
+## 30.0.1
+
+### Patch Changes
+
+- e343d22: Version DB to 47
+- f9b8c07: Bump registry version to v11.2
+
 ## 30.0.0
 
 ### Patch Changes